+++ data/tasks.txt
@@ -1,9 +1,6 @@
-[D][ ] die (by: 12/12/12)
-[D][ ] die (by: 12 Dec 0012 20:12)
-[D][ ] die1 (by: 12/12/12 12:12)
-[D][ ] die2 (by: 13 Dec 0012 07:59)
-[D][ ] die3 (by: 13 Dec 2012 07:59)
-[D][ ] die5 (by: 12 Dec 2012 16:00)
-[D][ ] die6 (by: 12 Dec 2012 08:00)
-[E][ ] die7 (from: 12 Dec 2012 08:00 to: 12 Dec 2012 09:00)
-[D][ ] afterlife (by: 11 Nov 2011 23:59)
+[A][X] hi
+[A][X] you
+[A][X] crazy
+[T][ ] homework
+[D][X] singing competition (by: 12 Dec 2023 14:52)
+[E][X] choir competition (from: 02 Sep 2023 02:00 to: 02 Sep 2023 12:00)
