+++ src/main/java/TaskType.java
@@ -1,3 +1,3 @@
 public enum TaskType {
-    ADD, TODO, DEADLINE, EVENT, LIST, MARK, UNMARK, DELETE, BYE, ECHO, UNKNOWN
+    ADD, TODO, DEADLINE, EVENT, LIST, MARK, UNMARK, DELETE, BYE, ECHO, UNKNOWN, HELP
 }
